Falta los combox para terminar el update

+++ app/Http/Controllers/UsuariosController.php
@@ -894,6 +894,39 @@ class UsuariosController extends Controller
                 }
         }
 
+        //conyuges
+        $usuario->conyuges()->delete($id);
+        if(isset($request->nombres_coy))
+        {
+                $nom=$request->nombres_coy;
+                $primero=$request->primero_coy;
+                $segundo=$request->segundo_coy;
+                $curp=$request->curp_coy;
+                $curp_carga=$request->carga_curp_coy;
+
+                //dd($curp_carga);
+
+                if($nom==""){$nom=0;}else{$nom=$request->nombres_coy;}
+                if($primero==""){$primero=0;}else{$primero=$request->primero_coy;}
+                if($segundo==""){$segundo=0;}else{$segundo=$request->segundo_coy;}
+                if($curp==""){$curp=0;}else{$curp=$request->curp_coy;}
+                if($curp_carga==""){$curp_carga=0;}else{
+                $filenamewithExt =$curp_carga->getClientOriginalName();
+                $curp_carga=$request->carga_curp_coy->storeAs('CURPCONYUGES',$filenamewithExt);
+                }
+
+
+
+                $usuario->conyuges()->create([
+                    'nombres_coy'=>$nom,
+                    'primero_coy'=>$primero,
+                    'segundo_coy'=>$segundo,
+                    'curp_coy'=>$curp,
+                    'carga_curp_coy'=>$curp_carga
+
+                ]);
+         }
+
 
          //borramos dependientes y se vuelve a crear
          $usuario->Descensientes()->delete($id);
